# platform/frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "CatalogReviewPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CatalogReviewResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "next_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Version"
          }
        },
        "required": [
          "items",
          "next_version"
        ],
        "title": "CatalogReviewPage",
        "type": "object"
      },
      "CatalogReviewRequest": {
        "additionalProperties": false,
        "properties": {
          "evidence": {
            "maxLength": 32000,
            "minLength": 1,
            "title": "Evidence",
            "type": "string"
          },
          "expected_version": {
            "minimum": 1.0,
            "title": "Expected Version",
            "type": "integer"
          },
          "idempotency_key": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Idempotency Key",
            "type": "string"
          },
          "reason": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Reason",
            "type": "string"
          },
          "reviewer": {
            "maxLength": 256,
            "minLength": 1,
            "title": "Reviewer",
            "type": "string"
          },
          "state": {
            "enum": [
              "active",
              "withdrawn"
            ],
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "expected_version",
          "state",
          "reason",
          "reviewer",
          "evidence",
          "idempotency_key"
        ],
        "title": "CatalogReviewRequest",
        "type": "object"
      },
...
      "ErrorEnvelopeResponse": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetailResponse"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelopeResponse",
        "type": "object"
      },
...
      "ResultReviewPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ResultReviewResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "ResultReviewPage",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Anonymous trusted-intranet crash analysis control plane. There are intentionally no login, RBAC, or DELETE endpoints.",
    "title": "Crash-Cap API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v3/symbol-catalog/pairs/{pair_id}/reviews": {
      "get": {
        "operationId": "list_pair_reviews_api_v3_symbol_catalog_pairs__pair_id__reviews_get",
        "parameters": [
          {
            "in": "path",
            "name": "pair_id",
            "required": true,
            "schema": {
              "title": "Pair Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "before_version",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Before Version"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogReviewPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          },
          "410": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          }
        },
        "summary": "List Pair Reviews"
      },
      "post": {
        "operationId": "submit_pair_review_api_v3_symbol_catalog_pairs__pair_id__reviews_post",
        "parameters": [
          {
            "in": "path",
            "name": "pair_id",
            "required": true,
            "schema": {
              "title": "Pair Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CatalogReviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogReviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          },
          "410": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          }
        },
        "summary": "Submit Pair Review"
      }
    },
    "/api/v3/symbol-catalog/pairs/{pair_id}/reviews/{review_id}/evidence": {
      "get": {
        "operationId": "get_review_evidence_api_v3_symbol_catalog_pairs__pair_id__reviews__review_id__evidence_get",
        "parameters": [
          {
            "in": "path",
            "name": "pair_id",
            "required": true,
            "schema": {
              "title": "Pair Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "review_id",
            "required": true,
            "schema": {
              "title": "Review Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogReviewEvidence"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          },
          "410": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          }
        },
        "summary": "Get Review Evidence"
      }
    },
...
    "/api/v3/workspaces/{workspace_id}/occurrences/{occurrence_id}/result-reviews": {
      "get": {
        "operationId": "list_result_reviews_api_v3_workspaces__workspace_id__occurrences__occurrence_id__result_reviews_get",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "title": "Workspace Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "occurrence_id",
            "required": true,
            "schema": {
              "title": "Occurrence Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultReviewPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          },
          "410": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          }
        },
        "summary": "List Result Reviews"
      },
      "post": {
        "operationId": "submit_result_review_api_v3_workspaces__workspace_id__occurrences__occurrence_id__result_reviews_post",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "title": "Workspace Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "occurrence_id",
            "required": true,
            "schema": {
              "title": "Occurrence Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResultReviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultReviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          },
          "410": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          }
        },
        "summary": "Submit Result Review"
      }
    },
    "/api/v3/workspaces/{workspace_id}/occurrences/{occurrence_id}/result-reviews/{review_id}": {
      "get": {
        "operationId": "get_result_review_api_v3_workspaces__workspace_id__occurrences__occurrence_id__result_reviews__review_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "title": "Workspace Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "occurrence_id",
            "required": true,
            "schema": {
              "title": "Occurrence Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "review_id",
            "required": true,
            "schema": {
              "title": "Review Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultReviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          },
          "410": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          }
        },
        "summary": "Get Result Review"
      }
    },
...
  }
}